+++ src/App.js
@@ -7,7 +7,7 @@ function App () {
   return (
     <>
       <Editor />
-      {/* <KeyMapTable table={suggestionMap} /> */}
+      <KeyMapTable table={suggestionMap} />
     </>
   )
 }
