+++ src/editor/plugins/suggestionMap.js
@@ -1,4 +1,4 @@
-export default {
+const map = {
   AMI: 'Acute Myocardial Infarction',
   CHD: 'Congenital Heart Disease',
   CHF: 'Congenital Heart Failure',
@@ -66,3 +66,25 @@ export default {
   OD: ' Oculus Dexter',
   'I&O': 'Intake and output'
 }
+
+const obj = {}
+const sa = {}
+Object.keys(map).forEach(
+  _key => {
+    const key = _key.toLowerCase()
+
+    obj[key] = [map[_key]]
+    if (!sa[key[0]]) {
+      sa[key[0]] = []
+    }
+    sa[key[0]].push(map[_key])
+  }
+)
+
+const suggestionMap = {
+  lt: ['left', 'light', 'lot'],
+  ...obj,
+  ...sa
+}
+
+export default suggestionMap
